Split table creation
